Flask_Document_Converter/requirements.txt
aspose-words==22.4.0
click==8.1.2
Flask==2.1.1
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
pip==22.0.4
PyPDF2==1.27.5
setuptools==61.0.0
Werkzeug==2.1.1
wheel==0.37.1
